+++ app/modules/sagas.ts
@@ -10,7 +10,6 @@ import { kycSagas } from "./kyc/sagas";
 import { settingsSagas } from "./settings/sagas";
 import { userAgentSagas } from "./userAgent/sagas";
 import { lightWalletSagas } from "./wallet-selector/light-wizard/sagas";
-import { walletSelectorSagas } from "./wallet-selector/sagas";
 import { web3Sagas } from "./web3/sagas";
 
 /**
@@ -20,7 +19,6 @@ function* allSagas(): Iterator<effects.Effect> {
   yield effects.all([
     effects.fork(kycSagas),
     effects.fork(initSagas),
-    effects.fork(walletSelectorSagas),
     effects.fork(dashboardSagas),
     effects.fork(settingsSagas),
     effects.fork(web3Sagas),
